requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.1
tweepy==4.14.0
feedparser==6.0.10
jinja2==3.1.2
pandas==2.1.1
nltk==3.8.1
python-dateutil==2.8.2 